+++ packages/playground/src/index.tsx
@@ -27,12 +27,12 @@
 import { ComponentsProvider } from '@looker/components'
 import React from 'react'
 import { render } from 'react-dom'
-import { TooltipDemo } from './Tooltip/TooltipDemo'
+import { ListLayoutDemo } from './Select/ListLayoutDemo'
 
 const App: React.FC = () => {
   return (
     <ComponentsProvider>
-      <TooltipDemo />
+      <ListLayoutDemo />
     </ComponentsProvider>
   )
 }